corrective-rag/requirements.txt
gitpython
langchain
langchain-community
langchain-openai
chromadb
openai
python-dotenv
unstructured
faiss-cpu
libmagic
bs4
tiktoken
streamlit
requests
pandas
numpy
pydantic
pymongo
